+++ lib/features/restaurant_tour/repositories/yelp_repository.dart
@@ -1,8 +1,9 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
-import 'package:restaurantour/models/restaurant.dart';
+import 'package:restaurantour/features/restaurant_tour/models/restaurant.dart';
 
-const _apiKey = '<PUT YOUR API KEY HERE>';
+const _apiKey =
+    'XwtYYNAyjLinimPxNBejp7RQrwpmoNJROqcWNe9ATgdLTAUxaWjGOMRXJUTTVknNZAKPtWQJ7T--axaZAIyWWwEKEuTwhTxK_OZPXF76inU4isZVBBQgVwWfWw3MZXYx';
 
 class YelpRepository {
   late Dio dio;
